+++ routes/insline.js
@@ -1,0 +1,15 @@
+/**
+ * Hears
+ * =====================
+ * If i write: @ptkdev_bot text in groups or chats
+ *
+ * @author:     Patryk Rzucidlo [@ptkdev] <info@ptkdev.it> (https://ptkdev.it)
+ * @license:    This code and contributions have 'GNU General Public License v3'
+ * @version:    0.1
+ * @link:       https://github.com/telegraf/telegraf
+ * @changelog:  0.1 initial release
+ *
+ */
+module.exports = function (bot, config, request) {
+
+};