+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.2.0 BETA (November 4, 2017)
+
+* Moved to javascript E6
+* Available webook and polling
+* Email Feature with nodemailer
+
 # v0.1.0 (November 2, 2017)
 
 * First Release.